+++ packages/vm/src/errors/vm_error.rs
@@ -86,6 +86,14 @@ pub enum VmError {
     WriteAccessDenied { backtrace: BT },
     #[error("Maximum call depth exceeded.")]
     MaxCallDepthExceeded { backtrace: BT },
+    #[error(
+        "The called function args arity does not match. The contract's method arity: {}",
+        contract_method_arity
+    )]
+    FunctionArityMismatch {
+        contract_method_arity: usize,
+        backtrace: BT,
+    },
 }
 
 impl VmError {
@@ -242,6 +250,13 @@ impl VmError {
             backtrace: BT::capture(),
         }
     }
+
+    pub(crate) fn function_arity_mismatch(contract_method_arity: usize) -> Self {
+        VmError::FunctionArityMismatch {
+            contract_method_arity,
+            backtrace: BT::capture(),
+        }
+    }
 }
 
 impl_from_err!(CommunicationError, VmError, VmError::CommunicationErr);
@@ -297,7 +312,7 @@ impl From<wasmer::RuntimeError> for VmError {
         let message = format!("RuntimeError: {}", original.message());
         debug_assert!(
             original.to_string().starts_with(&message),
-            "The error message we created is not a prefix of the error message from Wasmer. Our message: '{}'. Wasmer messsage: '{}'",
+            "The error message we created is not a prefix of the error message from Wasmer. Our message: '{}'. Wasmer message: '{}'",
             &message,
             original
         );
@@ -379,7 +394,7 @@ mod tests {
     }
 
     #[test]
-    fn cyrpto_err_works() {
+    fn crypto_err_works() {
         let error = VmError::crypto_err(CryptoError::generic_err("something went wrong"));
         match error {
             VmError::CryptoErr {